first try

+++ tip_calculator.py
@@ -4,9 +4,9 @@ cost = int(input('What is the cost of your meal? '))
 tip = int(input('What percentage tip would you like to give? '))
 split = int(input('How many of you to split the bill? '))
 
-your_tip = float(tip / 100 * cost)
-total_bill = float(your_tip + cost)
-split_bill = float(cost / split)
+your_tip = (tip / 100 * cost)
+total_bill = (your_tip + cost)
+split_bill = (cost / split)
 
 print(f'Welcome, {name}')
 print(f'You should leave a tip of: ${your_tip}')
